Added function to split entry into lines
With function, entry is split into 17 character lines with intelligent wrapping using
hyphens. With this splitting method an entry should in theory be able to have between
107 characters (worst case), and 125 characters (best case).

Worst case here is each line ending with a space as the 16th 'character', as spaces at
the end of a line can be considered free. This leads to two empty spaces at the end of
each of the first 6 lines for the worst case, as there's no reasonable way to hyphenate
a word at the end of a line with a space as character 16.

Best case here is each line ending with an actual last character of a word as the 17th
character on each line. This saves the need to include a space between that word and the
next, saving 1 space each on the first 6 lines, leading to a max of 125 characters.