Quick image color test
After discovering the gray by using color 2 (red) in black/white mode
I'm curious if it's necessary that it's actually red in the color map
Or if only the number matters. Color-test is using a yellow instead,
This would be nicer for me to work with just due to it more accurately
Representing what my own phat will display. Hopefully just the fact that
the pixels are still color 2 will do it