Modified remaining sprites
Footprint added, and blank space to fit sprites for other game version
If a pokemon has multiple forms the space will be added horizontally
to the right of the base. The default game version will be 1, using
the lower sprite of the two by default, and the default form is 0,
the furthest left sprite.