Porting and documentation
Ported several more methods to the dex-entry
file and thoroughly documented input values and functionality.

Functionality of the methods themselves were unchanged, though arguments 
for the image and fontsheet to be used were added.

Greenscreen was edited to hit all of the
important images in the new assets folder,
mockup folder is ignored because it should
consist of only full images, none of which
should need transparency properly set.