Initial corrections to numeric display
Numeric displays width argument is now actually a width
Previously it was treated more as the right bound of the
display rectangle for the numeric data, for the current
display area that happens to work out ok, but if I want
to rearrange the entry display later or reuse the numeric
display routine elsewhere it doesn't work in my favor.

Width argument is now properly a width and a right bound
is calculated quickly in the routine, simple change now
in exchange for reducing headache later.