MERGE MAKES NO SENSE
I have no idea what atom did, I'm going to blame it on atom in the end. 
"Oh you have the HEAD still sitting there, but we won't show it to you 
for you to actually fix it, and if you try to stage changes we're going 
to make it vanish instead of staging"

Whatever, it should be fixed now.