Documented fontsheet metadata intentions.
Made some changes to planned chunk usage in each fontsheet
The thought of using zTXt has passed, realized that iTXt is
essentially mandatory as the Latin-1 character set required by
the tEXt and zTXt is overly restrictive for use in storing
the fontsheet string data.

Additionally, made provision in documentation for including
the width of the fontsheet in characters. This could also be
calculated by the width of the image and the width of each
character, so this may change. Including a character width
of the sheet does allow for adding other image data off the
right of the font image data rather than just below, so it
is likely to become and remain a value kept in a chunk within
the image itself.

Still to do is writing a short script or tool to more easily
add this chunk data to each fontsheet.