Slight modifications to dex-entry, more notes
Dex-entry is almost to making use of the metadata pulled
from the chunk data. The current textual display methods
need to be refactored to more cleanly support the number
of variables being passed in to account for font variations.

Greenscreen script has been discovered to completely clobber
all the useful ancillary chunk data that had already been
added to the image. This is *probably* my fault, for the
naive import/export that it does, but it's still an issue
that needs to be corrected.