Began work on font class
It became drastically apparent after implementing metadata into the font 
images that the codebase was becoming too complex. Argument counts on 
some functions were ballooning out of control and actions needed to be 
taken to fix the problem.

The first of these actions is the creation of the Font class. This class 
will be initialized with the filename for the font of choice, and will 
extract and hold all the important values and metadata within itself. 
More functions are likely to be shifted to the font class itself as time 
goes on.