Functional font class, more to refactor yet
Font class holds metadata and the image for the
font sheets. Main public method currently is to
fetch a chartacter from the sheet.

Debating whether or not to split out paste
methods for line/lines to a different class
or at least a different file, definitely not
sure which would be the best option for
maintainability, but I don't think leaving them
as methods within the main file is the best.

Need to fiddle with a method to display floats,
and moving the kludge fix to move the numbers
closer to the decimal into that method so its
much more readable for maintenance.