Now using font class, image output saved
Finally got around to actually putting the font class to use instead of 
the old method calls, so was able to adjust some functions to account 
for that. There may still be more adjustments and cleanups that can be 
made but this is a start.

Consolidated the location values in each function call to an XY tuple. 
This doesn't strictly simplify the initial call necessarily but can 
simplify later calls within functions.

Black has been adopted for formatting of files so some of that has been 
changed as well.

Next to do is either implement logging or to clean up entry and numeric 
display functions to have a few less magic numbers sprinkled around.