Rudimentary testing UI
Wrote a basic GUI with PyQT to help expediate testing while developing 
on my desktop, can now just click a button to update the output instead 
of rerunning script in console and loading image.

I'd like to expand this GUI's functionality in the future but that's a 
project for a completely different day when everything else is working.