Pokemon class now contains its own sprite data
The pokemon class contains all of its loaded sprites in a list and the 
footprint contained separately.
Loading these is currently hardcoded but I would like to change this in 
the future potentially to accomodate for differing spritesheet sizes 
noted in the image text tags.
The dex-entry routines now also randomly choose a sprite from this list 
of sprites.

Minor bugfixes in logging configuration and test_ui files.