Numerics from font class
get_numeric returns an image of a numeral with optional prefix and 
suffix. This is an effort to clean the nightmare of the old magic of 
display_numeric in dex-entry, but doesn't look as nice just yet.

Small bugfix in poke class file loading

Moved create_mask to a utility file for use in other files