Image Fixes, Greenscreen Fix
Greenscreen script no longer clobbers text tags but may be slow as for 
each image it's reading and writing the entire image twice. This can 
probably (almost certainly) be avoided if I looked into the raw png data 
and modifying it directly, but given this script shouldn't run very 
often it's not a high priority.