Fix of a font sheet size, thoughts on DB
When I was fiddling with some of the font sheets the yesterday I somehow 
managed to grow the large gscfont from 3kb to 20kb with no visual 
difference. Pulling up the chunks revealed that gimp converting the 
color profile had added a significantly large chunk or two to the file 
which I manually removed. The interesting thing is removing this brought 
the size down significantly two around a third the smaller gscfont, 
which leads me to think it may be interesting to look into optimizing 
the other font files and sprites in the future to minimize space usage, 
though this is next to the lowest priority currently.

I also spent a lot of time considering how I want to build the database 
to go with what I'd like this program to support eventually, and decided 
on a few key things. First and most importantly, I'm currently thinking 
of restricting the pokemon range to those available within the 
generation 1/2 games, at least for the forseeable future. A big part of 
this stems from using modified sprites from the gen 2 games which are 
the perfect size for this. Also, because of this I intend to fill the 
database with information specifically from the gen2 games, with perhaps 
dex entries being the only exception (provided when wrapped they would 
be short enough for the screen). I used dbdiagram.io to mock up the 
start of my current plan for the db schema, just need to write it out in 
sqlalchemy.

Next steps are to build the database (likely sqlite for ease of use but 
may have to change in the future, this is a learning project), then 
populate a few entries by hand to test linking it up to the dex-entry 
code, and then finally write a scraper to read bulbapedia or serebii for 
all the actual data necessary.