Schema written, default entries, loading
Wrote the schema for the pokemon db for this application and after much 
trial and tribulation it is now properly built up when running 
pokeschema.py standalone. I tried to make sure that it wouldn't need 
much changing in the future but I'm sure I've missed something that I'll 
have to deal with later.

Pokedefaults.py generates all the default rows for the tables that are 
replacing enums and adds them to the database as sqlite doesn't support 
enums.

I removed the notes file as it's no longer necessary, and the pokemon 
class now loads its data from a database connection. The only problem 
here is that I need to make the connection still, and more importantly 
need to populate the database which is going to be a doozy.